requirements.txt
kafka-python
pandas
python-dotenv
torch
joblib
numpy
scikit-learn
shap
